server/requirements.txt
together
pillow
requests
flask
flask-cors
easyocr
opencv-python
torch
PyMuPDF
dotenv
numpy